+++ src/User/UserAttributeService.php
@@ -6,15 +6,15 @@ namespace Dbp\Relay\CoreBundle\User;
 
 use Dbp\Relay\CoreBundle\API\UserSessionInterface;
 
+/**
+ * @internal
+ */
 class UserAttributeService
 {
-    private UserAttributeMuxer $userAttributeMuxer;
-    private UserSessionInterface $currentUserSession;
-
-    public function __construct(UserSessionInterface $currentUserSession, UserAttributeMuxer $userAttributeMuxer)
+    public function __construct(
+        private readonly UserSessionInterface $currentUserSession,
+        private readonly UserAttributeMuxer $userAttributeMuxer)
     {
-        $this->userAttributeMuxer = $userAttributeMuxer;
-        $this->currentUserSession = $currentUserSession;
     }
 
     public function getCurrentUserIdentifier(): ?string
@@ -40,18 +40,6 @@ class UserAttributeService
      */
     public function getCurrentUserAttribute(string $userAttributeName, mixed $defaultValue = null): mixed
     {
-        return $this->getUserAttribute($this->getCurrentUserIdentifier(), $userAttributeName, $defaultValue);
-    }
-
-    /**
-     * Gets a user attribute for a user.
-     *
-     * @param mixed|null $defaultValue The value to return if the user attribute is declared but not specified for the current user
-     *
-     * @throws UserAttributeException If the user attribute is undeclared
-     */
-    public function getUserAttribute(?string $userIdentifier, string $userAttributeName, mixed $defaultValue = null): mixed
-    {
-        return $this->userAttributeMuxer->getAttribute($userIdentifier, $userAttributeName, $defaultValue);
+        return $this->userAttributeMuxer->getAttribute($this->getCurrentUserIdentifier(), $userAttributeName, $defaultValue);
     }
 }